abstract out rule editor bottom sheet
